Simplify, normalize px
